Convert IIIFViewer
