+++ cardigan/stories/components/IIIFViewer/IIIFViewer.stories.mdx
@@ -1,0 +1,16 @@
+import { Meta, Canvas, Story, ArgsTable, Description } from '@storybook/addon-docs/blocks';
+import IIIFViewer from '@weco/common/views/components/IIIFViewer/IIIFViewer';
+import Readme from '@weco/common/views/components/IIIFViewer/README.md';
+import * as stories from './IIIFViewer.stories.tsx';
+
+<Meta title={`Components/IIIFViewer`} component={IIIFViewer} />
+
+# IIIFViewer
+
+<Description>{Readme}</Description>
+
+<Canvas>
+  <Story story={stories.basic} inline={false} height={'100vh'} />
+</Canvas>
+
+<ArgsTable />
